update requirements (#12)
* Change order of imports in conftest.py

isort showed an error that the imports were not sorted
correctly. This commit changes the sorting so that
isort does not throw an error anymore.

* change requirements of isort

Starting with isort 4.3 some imports are sorted differently.
Raising the required version of isort makes sure that imports
are always sorted correctly. (changed to the same version
invenio requires).

* change requirement of invenio-oauth2server

Raised requirement of invenio-oauth2server to the
most recent one (1.0.3) to take care of a version
conflict with oauthlib.

* change requirement of invenio-accounts

Lowest requirements were raising a version conflict since
invenio-oauth2server needs >=1.0.0. Requirement was raised
to the same version.

* update invenio dependencies to released versions where possible

* added webargs==4.4.1 requirement temporarily

tests were failing because webargs 5 requires simplejson to be
installed on Python 2. This is currently not happening in
invenio-files-rest. Can be removed once invenio-files-rest is updated.